feat: reorganize server.py
