+++ mcp_server/tools/auth.py
@@ -1,0 +1,25 @@
+"""Authentication implementation logic."""
+import requests
+from typing import Dict, Any
+
+
+def tidal_login(flask_app_url: str) -> Dict[str, Any]:
+    """Implementation logic for TIDAL login."""
+    try:
+        # Call your Flask endpoint for TIDAL authentication
+        response = requests.get(f"{flask_app_url}/api/auth/login")
+
+        # Check if the request was successful
+        if response.status_code == 200:
+            return response.json()
+        else:
+            error_data = response.json()
+            return {
+                "status": "error",
+                "message": f"Authentication failed: {error_data.get('message', 'Unknown error')}"
+            }
+    except Exception as e:
+        return {
+            "status": "error",
+            "message": f"Failed to connect to TIDAL authentication service: {str(e)}"
+        }
